feat: move note/elements composables to commons (batch 18)
Move portable composables from ui/note/elements/ to commons for KMP:

- DisplayPoW: moved to commons (pure compose, no Android deps)
- BoostedMark: moved to commons with String param (Android wrapper resolves R.string)
- getCommunityShortName: moved utility function to commons
- AddRemoveButtons: delegates to existing commons ActionButtons

Added theme extensions to commons ThemeExtensions.kt:
- placeholderText, lessImportantLink color extensions
- Font14SP, HalfStartPadding, HalfTopPadding, ButtonBorder

Original Android files retained as thin wrappers/previews.

Part of the KMP iOS migration tracked in #2238.
